refactor: centralize scattered utility functions into lib/utils/
Extract 9 helper functions from 7 files into dedicated utility modules:

- lib/utils/format.ts: formatBytes (deduplicated from 2 files),
  formatUptime, formatRelativeTime
- lib/utils/crypto.ts: generatePassword
- lib/utils/parse.ts: parseKeyValueString (née parseCredentials)
- lib/utils/errors.ts: isNetworkError
- lib/utils/validation.ts: isSafeShellArg
- lib/utils/redis-lock.ts: acquireLock, releaseLock
- lib/proxmox/utils.ts: extractIpFromNet0

Document utility organization conventions in CLAUDE.md with
placement rules and decision framework for future helpers.
