+++ apps/dashboard/src/components/containers/container-card.tsx
@@ -6,14 +6,7 @@ import { StatusBadge } from "./status-badge";
 import { ContainerActions } from "./container-actions";
 import type { ContainerWithStatus } from "@/lib/containers/data";
 import type { ServiceStatus } from "@/generated/prisma/client";
-
-function formatBytes(bytes: number): string {
-  if (bytes === 0) return "0 B";
-  const units = ["B", "KB", "MB", "GB", "TB"];
-  const i = Math.floor(Math.log(bytes) / Math.log(1024));
-  const value = bytes / Math.pow(1024, i);
-  return `${value.toFixed(value >= 100 ? 0 : 1)} ${units[i]}`;
-}
+import { formatBytes } from "@/lib/utils/format";
 
 /** Color dot for service status */
 function ServiceDot({ status }: { status: ServiceStatus }) {
