+++ apps/dashboard/src/components/containers/detail/overview-tab.tsx
@@ -22,31 +22,7 @@ import {
 import { Badge } from "@/components/ui/badge";
 import { Separator } from "@/components/ui/separator";
 import type { ContainerDetailData } from "@/lib/containers/data";
-
-// ============================================================================
-// Helpers
-// ============================================================================
-
-function formatBytes(bytes: number): string {
-  if (bytes === 0) return "0 B";
-  const units = ["B", "KB", "MB", "GB", "TB"];
-  const i = Math.floor(Math.log(bytes) / Math.log(1024));
-  const value = bytes / Math.pow(1024, i);
-  return `${value.toFixed(i > 1 ? 1 : 0)} ${units[i]}`;
-}
-
-function formatUptime(seconds: number): string {
-  if (seconds === 0) return "N/A";
-  const days = Math.floor(seconds / 86400);
-  const hours = Math.floor((seconds % 86400) / 3600);
-  const minutes = Math.floor((seconds % 3600) / 60);
-
-  const parts: string[] = [];
-  if (days > 0) parts.push(`${days}d`);
-  if (hours > 0) parts.push(`${hours}h`);
-  if (minutes > 0) parts.push(`${minutes}m`);
-  return parts.length > 0 ? parts.join(" ") : "< 1m";
-}
+import { formatBytes, formatUptime } from "@/lib/utils/format";
 
 // ============================================================================
 // Types
