+++ apps/dashboard/src/lib/containers/monitoring.ts
@@ -3,22 +3,7 @@
 
 import { SSHSession, connectWithRetry } from "@/lib/ssh";
 import type { SSHExecResult } from "@/lib/ssh";
-
-// ============================================================================
-// Input Sanitization
-// ============================================================================
-
-/**
- * Validate that a string is safe for use in a shell command.
- * Allows only alphanumeric characters, hyphens, underscores, dots, and @.
- * This covers valid systemd unit names and filesystem paths without
- * shell metacharacters.
- */
-const SAFE_SHELL_ARG = /^[a-zA-Z0-9._@-]+$/;
-
-function isSafeShellArg(value: string): boolean {
-  return SAFE_SHELL_ARG.test(value);
-}
+import { isSafeShellArg } from "@/lib/utils/validation";
 
 // ============================================================================
 // Types
